Adds new Opinion Variants B & C; renames old variants
