+++ src/fronts/opinion/Opinion.tsx
@@ -3,6 +3,8 @@ import { Collection as ICollection } from "../../api";
 import { Collections } from "../../components/Collections";
 import { Collections as CollectionsVariantB } from "./variantB/Collections";
 import { Collections as CollectionsVariantC } from "./variantC/Collections";
+import { Collections as CollectionsVariantY } from "./variantY/Collections";
+import { Collections as CollectionsVariantZ } from "./variantZ/Collections";
 
 export const Opinion: React.FC<{
     frontId: string;
@@ -10,7 +12,27 @@ export const Opinion: React.FC<{
     salt: string;
     variant?: string;
 }> = ({ frontId, collections, salt, variant }) => {
-    if (variant === "c") {
+    if (variant === "z") {
+        // NOT IN USE
+        // FKA VARIANT C
+        return (
+            <CollectionsVariantZ
+                frontId={frontId}
+                collections={collections}
+                salt={salt}
+            ></CollectionsVariantZ>
+        );
+    } else if (variant === "y") {
+        // NOT IN USE
+        // FKA VARIANT B
+        return (
+            <CollectionsVariantY
+                frontId={frontId}
+                collections={collections}
+                salt={salt}
+            ></CollectionsVariantY>
+        );
+    } else if (variant === "c") {
         return (
             <CollectionsVariantC
                 frontId={frontId}
