+++ src/fronts/opinion/variantY/components/CommentCollection.tsx
