+++ src/fronts/opinion/variantY/components/DefaultCollection.tsx
