+++ src/fronts/opinion/variantY/components/EditorialCollection.tsx
