+++ src/fronts/opinion/variantY/components/Grid.tsx
