+++ src/fronts/opinion/variantY/components/LinkCollection.tsx
