+++ src/fronts/opinion/variantY/components/MediaCollection.tsx
