+++ src/fronts/opinion/variantZ/components/CommentCollection.tsx
