+++ src/fronts/opinion/variantZ/components/DefaultCollection.tsx
