+++ src/fronts/opinion/variantZ/components/EditorialCollection.tsx
