+++ src/fronts/opinion/variantZ/components/Grid.tsx
