+++ src/fronts/opinion/variantZ/components/LinkCollection.tsx
