+++ src/fronts/opinion/variantZ/components/MediaCollection.tsx
